+++ CHANGES.rst
@@ -6,6 +6,8 @@ For changes before version 3.0, see ``HISTORY.rst``.
 4.0a6 (unreleased)
 ------------------
 
+- Sanitize and test `RoleManager` role handling
+
 - Drop `Record` dependency, which now does its own security declaration.